changed lots of things

+++ distributions.h
@@ -3,6 +3,6 @@
 
 #include "fraction.h"
 double Beta_probability(double lower_bound, double upper_bound, double alpha1, double alpha2);
-std::vector<double> parameter_distributions(MC model, std::vector<double> lower_bounds, std::vector<double> upper_bounds);
+std::vector<unsigned> parameter_distributions(MC model, std::vector<double> lower_bounds, std::vector<double> upper_bounds);
 
 #endif
